api/requirements.txt
azure-functions
azure-identity
azure-cosmos